split uploading into a separate function
