+++ lib/app_web/controllers/api_controller.ex
@@ -9,16 +9,7 @@ defmodule AppWeb.ApiController do
 
     if fileIsAnImage do
 
-      # Create `CID` from file contents
-      {:ok, file_binary} = File.read(image.path)
-      file_cid = Cid.cid(file_binary)
-      file_name = "#{file_cid}.#{Enum.at(MIME.extensions(image.content_type), 0)}"
-
-      # Upload to S3
-      upload = image.path
-      |> ExAws.S3.Upload.stream_file()
-      |> ExAws.S3.upload("imgup-original", file_name, acl: :public_read)
-      |> ExAws.request(get_ex_aws_request_config_override())
+      upload = App.Upload.upload(image)
 
       # Check if upload was successful
       case upload do
@@ -45,8 +36,4 @@ defmodule AppWeb.ApiController do
   def create(conn, _params) do
     render(conn |> put_status(400), :field_error)
   end
-
-  def get_ex_aws_request_config_override,
-    do: Application.get_env(:ex_aws, :request_config_override)
-
 end
